♻️ Refactor: pagination code spliting (#34)
