+++ src/app/(auth)/profile/[userId]/_components/MyFeed.tsx
@@ -4,7 +4,7 @@ import { useQuery } from '@tanstack/react-query';
 import { getUserFeeds } from '@/api/profile';
 import { NoContent } from '@/components/atoms/noContent/NoContent';
 import { Loading } from '@/app/(auth)/_components/Loading';
-import { Pagination } from './Pagination';
+import { Pagination } from '@/components/molecules/pagination';
 type UserIdProps = {
   userId: number;
 };
