+++ src/app/(auth)/profile/[userId]/_components/ProfilePage.tsx
@@ -7,12 +7,7 @@ type Props = {
 };
 export const ProfilePage = async ({ userId }: Props) => {
   const currentUser = await getServerUserInfo();
-  // const feeds = await getMyfeeds();
-  // const act = await getMyActives();
-  // console.log('피드', feeds);
-  // console.log('액트', act);
-  // const user = await getUserFeeds(userId);
-  // console.log(user);
+
   if (currentUser && currentUser?.id == userId) {
     return <MyProfile userData={currentUser} isOwnProfile={true} />;
   }
