+++ src/app/(auth)/profile/[userId]/_components/Timeline.tsx
@@ -6,7 +6,7 @@ import { getMyActives } from '@/api/profile';
 import { useQuery } from '@tanstack/react-query';
 import { Loading } from '@/app/(auth)/_components/Loading';
 import { NoContent } from '@/components/atoms/noContent/NoContent';
-import { Pagination } from './Pagination';
+import { Pagination } from '@/components/molecules/pagination';
 
 export const Timeline = () => {
   const arr = activeDummyData;
@@ -26,7 +26,7 @@ export const Timeline = () => {
         const timeLine = simplifyTimeLineData(data.data[0]);
         return <UserTimeline key={timeLine.id} timeline={timeLine} />;
       })}
-      <Pagination />
+      {/* <Pagination /> */}
     </>
   ) : (
     <NoContent msg='내 활동이 없습니다.' />
