+++ src/components/atoms/noContent/index.ts
@@ -1,0 +1,1 @@
+export * from './NoContent';
