+++ src/components/atoms/pageNumberButton/index.ts
@@ -1,0 +1,1 @@
+export * from './PageNumberButton';
