+++ src/components/atoms/paginationButton/index.ts
@@ -1,0 +1,1 @@
+export * from './PaginationButton';
