+++ src/components/molecules/pagination/index.ts
@@ -1,0 +1,1 @@
+export * from './Pagination';
